requirements.txt
python-dotenv
python-telegram-bot==13.9
python-whois
unshortenit
tldextract
ipwhois